week03_fast_pipelines/homework/requirements.txt
datasets==2.9.0
imageio==2.25.0
jpeg4py==0.1.4
nvprof==0.2
opencv-python==4.7.0.68
scikit-image==0.19.3
pandas==1.5.3
py-spy==0.3.14
einops==0.7.0
torch==2.4.0
torchtext
torchvision==0.19.0
tqdm==4.64.1
transformers==4.48.2
vit_pytorch==0.40.2
gdown==4.7.3
matplotlib==3.8.2
pytest==8.3.5
wandb==0.19.8
hydra-core==1.3
